requirements.txt
python-dotenv
pandas
genanki
google-genai